+++ devel/0092.md
@@ -1,9 +1,10 @@
 # [0092] 拆分 s7_scheme_let.c 并在 (liii base) 导出 let 相关函数
 
 ## 任务相关的代码文件
-- src/s7.c （后续子任务：拆出 s7_scheme_let.c）
-- src/s7_scheme_let.c （后续子任务：新增）
-- src/s7_scheme_let.h （后续子任务：新增）
+- src/s7.c （已修改：迁出 let 相关函数，约 34+5 个函数去 static）
+- src/s7_scheme_let.c （已新增）
+- src/s7_scheme_let.h （已新增）
+- src/s7_internal.h （已修改：追加迁移所需的宏/枚举/声明）
 - goldfish/liii/base.scm （已修改：新增 20 个 let 相关函数的 export）
 - tests/liii/base/*-test.scm （已新增 20 个测试文件）
 - devel/0092.md
@@ -26,6 +27,46 @@ done
 # 完整测试套件
 bin/gf test --all
 ```
+
+## 2026-07-30 第二个子任务：从 s7.c 拆出 s7_scheme_let.c / s7_scheme_let.h
+
+### What
+1. 新增 `src/s7_scheme_let.c`（约 1350 行）与 `src/s7_scheme_let.h`（44 个函数声明），
+   从 `src/s7.c` 迁出 63 个代码区段：`g_unlet` 到 `g_is_defined` 的密集块，
+   外加 `g_funclet` 与 `g_owlet`，涵盖 unlet / openlet / coverlet / varlet /
+   cutlet / sublet / inlet / let->list / let-ref / let-set! / rootlet / curlet /
+   outlet / symbol->value / symbol->dynamic-value / defined? / funclet / owlet
+   及其私有辅助函数。`s7.c` 从 88327 行缩减到 87161 行。
+2. `src/s7_internal.h` 追加约 260 行：迁移代码所需的宏（含 S7_DEBUGGING /
+   WITH_GCC / REPORT_ROOTLET_REDEF 多分支变体）、`starlet_t` 枚举、
+   外部全局变量与函数声明。
+3. `s7.c` 中约 34 个 static 函数去 `static`（其中
+   `add_slot_at_end`/`add_slot_checked_with_id`/`add_slot_unchecked`/
+   `checked_slot_set_value`/`is_constant_symbol` 用 `extern inline` 保持内联），
+   另去 `static` 5 个仅 S7_DEBUGGING 下编译的调试函数
+   （`check_let_set_slots`/`set_opt2_1`/`set_opt3_1`/`gc_protect_2`/`unstack_1`）。
+4. `xmake.lua` 注册 `src/s7_scheme_let.c`；release 与 debug（-DS7_DEBUGGING）
+   两种模式均编译链接通过；`bin/gf test --all` 1492 项全部通过，
+   20 个 let 测试文件单独运行也全部通过。
+
+### Why
+`s7.c` 是 88k 行的单体文件，按既定路线（0074 起陆续拆出 symbol/inexact/
+continuation 等模块）继续拆分，let 相关函数是文档中规划好的下一块。
+
+### How
+- **边界判定靠引用计数**：用脚本统计每个候选函数被 s7.c 其余部分引用的次数，
+  求值热路径函数（lookup、position_of、is_proper_quote、starlet 等，被引用
+  数十到数百次）留在 s7.c，只迁出 20 个公开 let API 及其私有辅助。
+- **`extern inline` 模式**：热函数在 s7.c 保留 `extern inline` 定义、
+  在 s7_internal.h 放 `extern inline` 声明，TU 内仍内联、TU 间有外部符号。
+- **H_/Q_ 宏留在 s7.c**：文档/签名宏虽定义在函数体内但作用于文件，
+  s7.c 尾部的 defun 注册点仍在使用，迁移时只加注释标明函数已迁出。
+- **`symbol_set_id` 的双形态**：release 下是宏、debug 下是函数，头文件需
+  用 `#if S7_DEBUGGING` 镜像该结构，否则 release 链接报 undefined reference。
+- **`make_integer`/`make_real` 例外**：s7.c 的快路径宏依赖 static 的
+  `small_ints` 数组，s7_scheme_let.c 中直接
+  `#define make_integer(Sc, N) s7_make_integer(Sc, N)`（函数内部有同样的
+  小整数快路径），避免把 small_ints 暴露出去。
 
 ## 2026-07-29 第一个子任务：在 (liii base) 导出 20 个 let 相关函数并补齐文档与测试
 
@@ -66,10 +107,7 @@ bin/gf test --all
     `(with-let (unlet) ...)` 以访问被遮蔽的原始版本
 
 ## 后续子任务（待做）
-- 从 `src/s7.c` 拆出 `s7_scheme_let.c` / `s7_scheme_let.h`
-- 拆分边界：`g_unlet`（9606）到 `g_is_defined`（11471）的密集块，
-  外加 `g_funclet`（34666）和 `g_owlet`（40690）
-- 关键约束：`g_simple_inlet` / `g_sublet_curlet` / `g_cdr_let_ref` /
-  `g_rootlet_ref` / `g_unlet_ref` / `g_cdr_let_set` / `g_unlet_set` /
-  `g_starlet_set` 被 s7.c 内部通过 `fn_proc() ==` 比较 和 `sc->xxx` 缓存
-  槽位引用，外迁时需去 `static` 通过头文件暴露符号
+- （暂无；拆分已完成。遗留约束备查：`g_simple_inlet` / `g_sublet_curlet` /
+  `g_cdr_let_ref` / `g_rootlet_ref` / `g_unlet_ref` / `g_cdr_let_set` /
+  `g_unlet_set` / `g_starlet_set` 被 s7.c 内部通过 `fn_proc() ==` 比较和
+  `sc->xxx` 缓存槽位引用，已通过 s7_scheme_let.h 暴露符号）
